+++ pkg/iam/iam_actions.go
@@ -48,9 +48,6 @@ const (
 	ActionMembershipUpdate = "iam:membership:update"
 	ActionMembershipDelete = "iam:membership:delete"
 
-	// Membership role actions
-	ActionMembershipRoleSetOwner = "iam:membership-role:set-owner"
-
 	// Membership Profile actions
 	ActionMembershipProfileGet        = "iam:membership-profile:get"
 	ActionMembershipProfileList       = "iam:membership-profile:list"